# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend App for simple social Live Chat",
        "title": "chat app",
        "contact": {
            "name": "Oleksii Shevchenko",
            "url": "https://github.com/mrOShevchenko"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/v1/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search users by username with autocomplete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "searchQuery string for search by username, minimum 1 character, maximum 40 characters",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "type of search, default: 'user', available: 'user', 'friend', 'blacklist'",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order of search, default: 'asc', available: 'asc', 'desc'",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit of search, default: '10', available: '1-1000'",
                        "name": "searchLimit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "success operation"
                }
            }
        },
...
        "models.Device": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Chrome 90.0.4430.212 (Linux x86_64)"
                },
                "token": {
                    "type": "string",
                    "example": "c26BG3n3VJbI0i2H8aXZmG:APA91bGoG3iqJxidumkVRKCniXoA-QsYfSXpc6qhAiWcaIVtAUG9nNKsxoEkL8j4ZVezXVFzJDpIYS6JHtpcJ2af0686djcfKDltLqVLkuWFVHoEnz9NtKV9hgQmof7MURLYQsaokGfM"
                },
                "type": {
                    "description": "web, android, ios",
                    "type": "string",
                    "example": "web"
                },
                "userId": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "blockedUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Device"
                    }
                },
                "followedUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "image.png"
                },
                "isOnline": {
                    "type": "boolean",
                    "example": true
                },
                "lastActive": {
                    "type": "integer",
                    "example": 1230000000
                },
                "username": {
                    "type": "string",
                    "example": "username"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Should contain Access JWT Token, with the Bearer started",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}